Update nodejs-ci.yml
Since I will need to run the acceptance tests in nodejs-cd.yml I think it makes sense to add all the files to the artifact not just the compiled (transpiled) binaries. This will ensure we are running the correct acceptance tests. Even though we will have the source files in the artifact we will NOT re-compile.