Update create-release.yml
Come up with a better way to store Releases. We will just store them in the artifact repository and have a GitHub pages page where they can be downloaded. As for the release artifact being deleted after 90 days we can handle this by creating a workflow that runs every so often that refreshes the release artifacts so they won't expire.